requirements.txt
cement==3.0.2
jinja2
pyyaml
colorlog
PyInquirer
pyfiglet
colorama
progress
asciimatics
v20
python-dateutil
asciiplotlib
numpy
